requirements.txt
dnspython>=2.6,<3.0
